Modularise Csvw classes
