+++ lib/csvlint.rb
@@ -15,11 +15,13 @@ require 'csvlint/error_collector'
 require 'csvlint/validate'
 require 'csvlint/wrapped_io'
 require 'csvlint/field'
-require 'csvlint/csvw_metadata_error'
-require 'csvlint/csvw_number_format'
-require 'csvlint/csvw_date_format'
-require 'csvlint/csvw_property_checker'
-require 'csvlint/csvw_column'
-require 'csvlint/csvw_table'
-require 'csvlint/csvw_table_group'
+
+require 'csvlint/csvw/metadata_error'
+require 'csvlint/csvw/number_format'
+require 'csvlint/csvw/date_format'
+require 'csvlint/csvw/property_checker'
+require 'csvlint/csvw/column'
+require 'csvlint/csvw/table'
+require 'csvlint/csvw/table_group'
+
 require 'csvlint/schema'
