--- lib/csvlint/csvw_metadata_error.rb
@@ -1,13 +1,0 @@
-module Csvlint
-
-  class CsvwMetadataError < StandardError
-
-    attr_reader :path
-
-    def initialize(path=nil)
-      @path = path
-    end
-
-  end
-
-end
