+++ spec/csvw_number_format_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe Csvlint::CsvwNumberFormat do
+describe Csvlint::Csvw::NumberFormat do
 
   it "should correctly parse #,##0.##" do
-    format = Csvlint::CsvwNumberFormat.new("#,##0.##")
+    format = Csvlint::Csvw::NumberFormat.new("#,##0.##")
     expect(format.pattern).to eq("#,##0.##")
     expect(format.prefix).to eq("")
     expect(format.numeric_part).to eq("#,##0.##")
@@ -16,42 +16,42 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should correctly parse ###0.#####" do
-    format = Csvlint::CsvwNumberFormat.new("###0.#####")
+    format = Csvlint::Csvw::NumberFormat.new("###0.#####")
     expect(format.primary_grouping_size).to eq(0)
     expect(format.secondary_grouping_size).to eq(0)
     expect(format.fractional_grouping_size).to eq(0)
   end
 
   it "should correctly parse ###0.0000#" do
-    format = Csvlint::CsvwNumberFormat.new("###0.0000#")
+    format = Csvlint::Csvw::NumberFormat.new("###0.0000#")
     expect(format.primary_grouping_size).to eq(0)
     expect(format.secondary_grouping_size).to eq(0)
     expect(format.fractional_grouping_size).to eq(0)
   end
 
   it "should correctly parse #,##,###,####" do
-    format = Csvlint::CsvwNumberFormat.new("#,##,###,####")
+    format = Csvlint::Csvw::NumberFormat.new("#,##,###,####")
     expect(format.primary_grouping_size).to eq(4)
     expect(format.secondary_grouping_size).to eq(3)
     expect(format.fractional_grouping_size).to eq(0)
   end
 
   it "should correctly parse #,##0.###,#" do
-    format = Csvlint::CsvwNumberFormat.new("#,##0.###,#")
+    format = Csvlint::Csvw::NumberFormat.new("#,##0.###,#")
     expect(format.primary_grouping_size).to eq(3)
     expect(format.secondary_grouping_size).to eq(3)
     expect(format.fractional_grouping_size).to eq(3)
   end
 
   it "should correctly parse #0.###E#0" do
-    format = Csvlint::CsvwNumberFormat.new("#0.###E#0")
+    format = Csvlint::Csvw::NumberFormat.new("#0.###E#0")
     expect(format.prefix).to eq("")
     expect(format.numeric_part).to eq("#0.###E#0")
     expect(format.suffix).to eq("")
   end
 
   it "should match numbers that match ##0 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("##0")
+    format = Csvlint::Csvw::NumberFormat.new("##0")
     expect(format.match("1")).to eq(true)
     expect(format.match("12")).to eq(true)
     expect(format.match("123")).to eq(true)
@@ -61,7 +61,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #,#00 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#,#00")
+    format = Csvlint::Csvw::NumberFormat.new("#,#00")
     expect(format.match("1")).to eq(false)
     expect(format.match("12")).to eq(true)
     expect(format.match("123")).to eq(true)
@@ -74,7 +74,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #,000 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#,000")
+    format = Csvlint::Csvw::NumberFormat.new("#,000")
     expect(format.match("1")).to eq(false)
     expect(format.match("12")).to eq(false)
     expect(format.match("123")).to eq(true)
@@ -87,7 +87,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #,##,#00 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#,##,#00")
+    format = Csvlint::Csvw::NumberFormat.new("#,##,#00")
     expect(format.match("1")).to eq(false)
     expect(format.match("12")).to eq(true)
     expect(format.match("123")).to eq(true)
@@ -100,7 +100,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #0.# correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.#")
+    format = Csvlint::Csvw::NumberFormat.new("#0.#")
     expect(format.match("1")).to eq(true)
     expect(format.match("12")).to eq(true)
     expect(format.match("12.3")).to eq(true)
@@ -110,7 +110,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #0.0 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.0")
+    format = Csvlint::Csvw::NumberFormat.new("#0.0")
     expect(format.match("1")).to eq(false)
     expect(format.match("12")).to eq(false)
     expect(format.match("12.3")).to eq(true)
@@ -120,7 +120,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #0.0# correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.0#")
+    format = Csvlint::Csvw::NumberFormat.new("#0.0#")
     expect(format.match("1")).to eq(false)
     expect(format.match("12")).to eq(false)
     expect(format.match("12.3")).to eq(true)
@@ -129,7 +129,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #0.0#,# correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.0#,#")
+    format = Csvlint::Csvw::NumberFormat.new("#0.0#,#")
     expect(format.match("1")).to eq(false)
     expect(format.match("12.3")).to eq(true)
     expect(format.match("12.34")).to eq(true)
@@ -141,7 +141,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should match numbers that match #0.###E#0 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.###E#0")
+    format = Csvlint::Csvw::NumberFormat.new("#0.###E#0")
     expect(format.match("1")).to eq(false)
     expect(format.match("12.3")).to eq(false)
     expect(format.match("12.34")).to eq(false)
@@ -151,7 +151,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match ##0 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("##0")
+    format = Csvlint::Csvw::NumberFormat.new("##0")
     expect(format.parse("1")).to eql(1)
     expect(format.parse("12")).to eql(12)
     expect(format.parse("123")).to eql(123)
@@ -161,7 +161,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #,#00 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#,#00")
+    format = Csvlint::Csvw::NumberFormat.new("#,#00")
     expect(format.parse("1")).to eq(nil)
     expect(format.parse("12")).to eql(12)
     expect(format.parse("123")).to eql(123)
@@ -174,7 +174,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #,000 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#,000")
+    format = Csvlint::Csvw::NumberFormat.new("#,000")
     expect(format.parse("1")).to eq(nil)
     expect(format.parse("12")).to eq(nil)
     expect(format.parse("123")).to eql(123)
@@ -187,7 +187,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #,##,#00 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#,##,#00")
+    format = Csvlint::Csvw::NumberFormat.new("#,##,#00")
     expect(format.parse("1")).to eq(nil)
     expect(format.parse("12")).to eql(12)
     expect(format.parse("123")).to eql(123)
@@ -200,7 +200,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #0.# correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.#")
+    format = Csvlint::Csvw::NumberFormat.new("#0.#")
     expect(format.parse("1")).to eql(1.0)
     expect(format.parse("12")).to eql(12.0)
     expect(format.parse("12.3")).to eql(12.3)
@@ -210,7 +210,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #0.0 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.0")
+    format = Csvlint::Csvw::NumberFormat.new("#0.0")
     expect(format.parse("1")).to eq(nil)
     expect(format.parse("12")).to eq(nil)
     expect(format.parse("12.3")).to eql(12.3)
@@ -220,7 +220,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #0.0# correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.0#")
+    format = Csvlint::Csvw::NumberFormat.new("#0.0#")
     expect(format.parse("1")).to eq(nil)
     expect(format.parse("12")).to eq(nil)
     expect(format.parse("12.3")).to eql(12.3)
@@ -229,7 +229,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #0.0#,# correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.0#,#")
+    format = Csvlint::Csvw::NumberFormat.new("#0.0#,#")
     expect(format.parse("1")).to eq(nil)
     expect(format.parse("12.3")).to eql(12.3)
     expect(format.parse("12.34")).to eql(12.34)
@@ -241,7 +241,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers that match #0.###E#0 correctly" do
-    format = Csvlint::CsvwNumberFormat.new("#0.###E#0")
+    format = Csvlint::Csvw::NumberFormat.new("#0.###E#0")
     expect(format.parse("1")).to eq(nil)
     expect(format.parse("12.3")).to eq(nil)
     expect(format.parse("12.34")).to eq(nil)
@@ -251,7 +251,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers normally when there is no pattern" do
-    format = Csvlint::CsvwNumberFormat.new()
+    format = Csvlint::Csvw::NumberFormat.new()
     expect(format.parse("1")).to eql(1)
     expect(format.parse("-1")).to eql(-1)
     expect(format.parse("12.3")).to eql(12.3)
@@ -269,7 +269,7 @@ describe Csvlint::CsvwNumberFormat do
   end
 
   it "should parse numbers including grouping separators when they are specified" do
-    format = Csvlint::CsvwNumberFormat.new(nil, ",")
+    format = Csvlint::Csvw::NumberFormat.new(nil, ",")
     expect(format.parse("1")).to eql(1)
     expect(format.parse("12.3")).to eql(12.3)
     expect(format.parse("12.34")).to eql(12.34)
